requirements/base.txt
# Core Python Dependencies
python-dotenv>=1.0.0,<2.0.0
requests>=2.28.0,<3.0.0
pyyaml>=5.4.0,<7.0.0

# Data Manipulation
pandas>=1.5.0,<3.0.0
numpy>=1.23.0,<2.0.0
sqlalchemy>=1.4.0,<3.0.0

# Frontend Framework
streamlit>=1.20.0,<2.0.0

# Visualization
plotly>=5.10.0,<6.0.0
matplotlib>=3.6.0,<4.0.0

# API Integration
fastapi>=0.95.0,<1.0.0
uvicorn>=0.22.0,<1.0.0

# Logging
loguru>=0.6.0,<1.0.0

# Database Connectivity
psycopg2-binary>=2.9.0,<3.0.0

# Cloud Platform
ibm-cloud-sdk-core>=3.15.0,<4.0.0
ibm-platform-services>=0.20.0,<1.0.0

# LLM and AI
ibm-watsonx-ai  # Confirmed working package

# JIRA and ALM Connectors
jira>=3.4.0,<4.0.0

# PDF and Document Processing
PyPDF2>=3.0.0,<4.0.0
mammoth>=1.4.0,<2.0.0

# Typing and validation
pydantic>=1.10.0,<3.0.0

# Additional Utilities
typing-extensions>=4.5.0,<5.0.0
setuptools>=65.0.0,<70.0.0

openpyxl
flask